Onom almost finished and working fine, bug fix of legs on jump

+++ source/entity/common_msgs.h
@@ -11,7 +11,7 @@ struct TMsgEntityCreated {
 
 struct TMsgOnomPet {
   VEC3 pos;
-  int type;
+  float type;
   DECL_MSG_ID();
 };
